Remove bitwarden::Kdf
