+++ crates/bitwarden-crypto/src/uniffi_support.rs
@@ -1,4 +1,20 @@
+use std::num::NonZeroU32;
+
 use crate::{CryptoError, EncString, UniffiCustomTypeConverter};
+
+uniffi::custom_type!(NonZeroU32, u32);
+
+impl UniffiCustomTypeConverter for NonZeroU32 {
+    type Builtin = u32;
+
+    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
+        Self::new(val).ok_or(CryptoError::ZeroNumber.into())
+    }
+
+    fn from_custom(obj: Self) -> Self::Builtin {
+        obj.get()
+    }
+}
 
 uniffi::custom_type!(EncString, String);
