+++ crates/bitwarden-uniffi/src/auth/mod.rs
@@ -1,10 +1,7 @@
 use std::sync::Arc;
 
-use bitwarden::{
-    auth::{password::MasterPasswordPolicyOptions, RegisterKeyResponse},
-    client::kdf::Kdf,
-};
-use bitwarden_crypto::HashPurpose;
+use bitwarden::auth::{password::MasterPasswordPolicyOptions, RegisterKeyResponse};
+use bitwarden_crypto::{HashPurpose, Kdf};
 
 use crate::{error::Result, Client};
