+++ crates/bitwarden/src/auth/login/mod.rs
@@ -45,3 +45,43 @@ pub(crate) async fn request_prelogin(
     let config = client.get_api_configurations().await;
     Ok(accounts_prelogin_post(&config.identity, Some(request_model)).await?)
 }
+
+#[cfg(feature = "internal")]
+use bitwarden_crypto::Kdf;
+
+#[cfg(feature = "internal")]
+pub(crate) fn parse_prelogin(response: PreloginResponseModel) -> Result<Kdf> {
+    use std::num::NonZeroU32;
+
+    use bitwarden_api_identity::models::KdfType;
+
+    use crate::util::{
+        default_argon2_iterations, default_argon2_memory, default_argon2_parallelism,
+        default_pbkdf2_iterations,
+    };
+
+    let kdf = response.kdf.ok_or("KDF not found")?;
+
+    Ok(match kdf {
+        KdfType::Variant0 => Kdf::PBKDF2 {
+            iterations: response
+                .kdf_iterations
+                .and_then(|e| NonZeroU32::new(e as u32))
+                .unwrap_or_else(default_pbkdf2_iterations),
+        },
+        KdfType::Variant1 => Kdf::Argon2id {
+            iterations: response
+                .kdf_iterations
+                .and_then(|e| NonZeroU32::new(e as u32))
+                .unwrap_or_else(default_argon2_iterations),
+            memory: response
+                .kdf_memory
+                .and_then(|e| NonZeroU32::new(e as u32))
+                .unwrap_or_else(default_argon2_memory),
+            parallelism: response
+                .kdf_parallelism
+                .and_then(|e| NonZeroU32::new(e as u32))
+                .unwrap_or_else(default_argon2_parallelism),
+        },
+    })
+}
