+++ crates/bitwarden/src/auth/mod.rs
@@ -15,7 +15,7 @@ use bitwarden_crypto::{HashPurpose, MasterKey};
 pub use register::{RegisterKeyResponse, RegisterRequest};
 
 #[cfg(feature = "internal")]
-use crate::{client::kdf::Kdf, error::Result};
+use crate::{client::Kdf, error::Result};
 
 #[cfg(feature = "internal")]
 async fn determine_password_hash(
@@ -24,7 +24,7 @@ async fn determine_password_hash(
     password: &str,
     purpose: HashPurpose,
 ) -> Result<String> {
-    let master_key = MasterKey::derive(password.as_bytes(), email.as_bytes(), &kdf.into())?;
+    let master_key = MasterKey::derive(password.as_bytes(), email.as_bytes(), kdf)?;
     Ok(master_key.derive_master_key_hash(password.as_bytes(), purpose)?)
 }
