+++ crates/bitwarden/src/error.rs
@@ -49,9 +49,6 @@ pub enum Error {
     #[error("The state file could not be read")]
     InvalidStateFile,
 
-    #[error("Number is zero")]
-    ZeroNumber,
-
     #[error("Internal error: {0}")]
     Internal(Cow<'static, str>),
 }
