+++ crates/bitwarden/src/platform/get_user_api_key.rs
@@ -50,7 +50,7 @@ fn get_secret_verification_request(
             .master_password
             .as_ref()
             .map(|p| {
-                let master_key = MasterKey::derive(p.as_bytes(), email.as_bytes(), &kdf.into())?;
+                let master_key = MasterKey::derive(p.as_bytes(), email.as_bytes(), kdf)?;
 
                 master_key.derive_master_key_hash(p.as_bytes(), HashPurpose::ServerAuthorization)
             })
