+++ crates/bitwarden/src/uniffi_support.rs
@@ -3,23 +3,10 @@ use std::{num::NonZeroU32, str::FromStr};
 use bitwarden_crypto::{AsymmEncString, EncString};
 use uuid::Uuid;
 
-use crate::{error::Error, UniffiCustomTypeConverter};
+use crate::UniffiCustomTypeConverter;
 
+uniffi::ffi_converter_forward!(NonZeroU32, bitwarden_crypto::UniFfiTag, crate::UniFfiTag);
 uniffi::ffi_converter_forward!(EncString, bitwarden_crypto::UniFfiTag, crate::UniFfiTag);
-
-uniffi::custom_type!(NonZeroU32, u32);
-
-impl UniffiCustomTypeConverter for NonZeroU32 {
-    type Builtin = u32;
-
-    fn into_custom(val: Self::Builtin) -> uniffi::Result<Self> {
-        Self::new(val).ok_or(Error::ZeroNumber.into())
-    }
-
-    fn from_custom(obj: Self) -> Self::Builtin {
-        obj.get()
-    }
-}
 
 uniffi::custom_type!(AsymmEncString, String);
